+++ lib/eva/lm/controller.py
@@ -842,7 +842,7 @@ def destroy_macro(m_id):
         raise ResourceNotFound
     try:
         i.destroy()
-        if eva.core.config.db_update == 1 and i.config_file_exists:
+        if eva.core.config.auto_save and i.config_file_exists:
             try:
                 eva.registry.key_delete(i.get_rkn())
             except:
@@ -901,7 +901,7 @@ def destroy_cycle(m_id):
     try:
         i.stop(wait=True)
         i.destroy()
-        if eva.core.config.db_update == 1 and i.config_file_exists:
+        if eva.core.config.auto_save and i.config_file_exists:
             try:
                 eva.registry.key_delete(i.get_rkn())
             except:
@@ -946,7 +946,7 @@ def destroy_dm_rule(r_id):
         i = dm_rules[r_id]
         i.destroy()
         DM.remove_rule(i)
-        if eva.core.config.db_update == 1 and i.config_file_exists:
+        if eva.core.config.auto_save and i.config_file_exists:
             try:
                 eva.registry.key_delete(i.get_rkn())
             except:
@@ -985,7 +985,7 @@ def destroy_job(r_id):
         i = jobs[r_id]
         i.unschedule()
         i.destroy()
-        if eva.core.config.db_update == 1 and i.config_file_exists:
+        if eva.core.config.auto_save and i.config_file_exists:
             try:
                 eva.registry.key_delete(i.get_rkn())
             except:
@@ -1108,7 +1108,7 @@ def remove_controller(controller_id):
     try:
         i = remote_ucs[_controller_id]
         i.destroy()
-        if eva.core.config.db_update == 1 and i.config_file_exists:
+        if eva.core.config.auto_save and i.config_file_exists:
             try:
                 eva.registry.key_delete(i.get_rkn())
             except:
@@ -1209,7 +1209,7 @@ def destroy_item(item):
         if not items_by_group[i.group]:
             del items_by_group[i.group]
         i.destroy()
-        if eva.core.config.db_update == 1 and i.config_file_exists:
+        if eva.core.config.auto_save and i.config_file_exists:
             try:
                 eva.registry.key_delete(i.get_rkn())
             except:
@@ -1312,8 +1312,8 @@ def connect_remote_controller(v):
 def stop():
     eva.core.plugins_exec('before_stop')
     # save modified items on exit, for db_update = 2 save() is called by core
-    if eva.core.config.db_update == 1:
-        save()
+    # if eva.core.config.db_update == 1:
+        # save()
     eva.lm.jobs.scheduler.stop()
     for i, v in cycles_by_id.copy().items():
         v.stop()